fix(plugins): keep prepared catalog types acyclic
